Improved compatibility with Apple's blurring appearance
These changes makes it easier for adding new controls that looks similar the native controls that support background blurring (like UIToolBar and UINavigationBar):
Added support for monochrome blurring: that's what's happening if setting the tintColor for UINavigationBar or UIToolBar -> the blurred image is converted to grayscale and then the tintColor is applied to it.
Removed the forced alpha value for the tintColor: for iOS > 7.0.4, Apple no longer ignores the alpha value for tintColor, so if you apply a tintColor with an alpha of value 1, you'd get a complete opaque view instead of semi-transparent blurred one.

Signed-off-by: Or Sharir <or@sharir.org>